requirements.txt
pydantic
earthengine-api
lightgbm
pytest
shapely
requests
rasterio
Pillow